Fix CLJ-931: Syntactically broken clojure.test/are tests succeed.
Fixed a test case in test/clojure/test_clojure/java_interop.clj only
because of the bug in clojure.test/are CLJ-931 is about.

No test case added that causes clojure.test/are to fail, because could
not determine how to write one that didn't also cause the Clojure
build tests to fail.

Signed-off-by: Stuart Halloway <stu@thinkrelevance.com>