CLJ-1325 - Warn on boxed math when *unchecked-math* is set.
When *unchecked-math* is set, static calls to any method in
clojure.lang.Numbers taking an Object or Number parameter will cause
a boxed math warning to be emitted. WarnBoxedMath is a new Java
annotation that can be used to explicitly mark methods as triggering
or not triggering the warning. Currently WarnBoxedMath is used to
mark a small number of methods that have Object parameters to not
trigger the warning.

Signed-off-by: Stuart Halloway <stu@cognitect.com>