CLJ-1380 - Ensure ex-info data argument is a map
The intention from the code and docs is that exceptions created by
ex-info should carry a map of additional data. The two-arg constructor
ensures that the data argument is a map but it was possible for data to
be nil if the three-arg constructor was used.

Apply the same argument checks from the two-arg ExceptionInfo
constructor to the three-arg constructor.

Signed-off-by: Stuart Halloway <stu@cognitect.com>