Clear reference to 'this' on all tail calls
Removes calls to emitClearLocals(), which is a no-op.

When the context is RETURN (indicating a tail call), and the operation
is an InvokeExpr, StaticMethodExpr, or InstanceMethodExpr, clear the
reference to 'this' which is in slot 0 of the locals.

Edge-case: Inside the body of a try block, we cannot clear 'this' at the tail
position as we might need to keep refs around for use in the catch or finally
clauses. Introduces another truthy dynamic binding var to track position being
inside a try block.

In a try block with no catch or finally, use enclosing context and return a
regular BodyExpr.

Adds two helpers to emitClearThis and inTailCall.

Adds test for the original reducer case and some try/catch cases.

Signed-off-by: Stuart Halloway <stu@cognitect.com>